Extract deprecated serializers
